requirements.txt
flask==3.0.0
flask-socketio==5.3.6
redis==5.0.1
scikit-learn==1.3.2
PyJWT==2.8.0
watchdog==3.0.0
prometheus-client==0.19.0
supabase==2.3.0
groq==0.4.1
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.2.1
